requirements.txt
pyttsx3
SpeechRecognition
requests
python-dotenv
googletrans==4.0.0-rc1
